requirements.txt
fastapi==0.103.1
uvicorn==0.23.0
sqlalchemy==2.0.20
pymysql==1.1.0
bcrypt 
pyjwt
passlib
jose   
huggingface_hub
json
python-dotenv